Rename context.rs into rules.rs
