+++ crates/analyzer/src/infer.rs
@@ -1,7 +1,7 @@
 //! Database for type inference.
 mod constraint;
-mod context;
 mod lower;
+mod rules;
 mod trees;
 
 #[cfg(test)]
@@ -22,6 +22,6 @@ pub trait InferDatabase: ScopeDatabase + SugarDatabase {
     #[salsa::interned]
     fn intern_type(&self, t: Type) -> TypeId;
 
-    #[salsa::invoke(context::infer_binding_group_query)]
+    #[salsa::invoke(rules::infer_binding_group_query)]
     fn infer_binding_group(&self, id: InFile<BindingGroupId>) -> Arc<InferBindingGroup>;
 }
