+++ crates/analyzer/src/infer/rules.rs
